Return new directory listing format on BaseClient#getListing
This adds support for the new directory listing format in the -02
specification of remoteStorage. It changes the `getListing` return value
to an object, which means apps using the old output need to be updated.

See the function documentation and unit test for details of how it
works and what changed exactly.