Fix params extraction from location hash
The current code broke applications that use hash navigation. For example, if I access my application with url http://toto.org/#settings, #settings is extracted by extractParams and the hash cleared, so my application never get it.
So I think that if hash is not of the form key=val&…, we should ignore it.