Update dropbox.js
On saving a file with non-English characters at its path, it creates a copy of the file with a url-encoded name. After some debugging I found out that response.path !== path (line 380) was causing the problem. So I bypassed the cleanPath() function by creating a new variable to store path value before it is converted.

Also I could not delete a file with non-English characters at its path. This is caused probably by the fact that on line 412 the path that was already converted by the cleanPath function, was again passed to encodeURIComponent function. So again a temp variable is created to store the path before conversion. 

This version has a more meaningful variable name (I hope!) : pathTempBeforeClean, pathTempBeforeCleanD