Fix RemoteStorage global missing in AMD build
We're accessing properties from the RemoteStorage global variable in
various places of the library. However, when `module.exports` was
available, it would only export the object, but not set the global.

Mid-term, we need to refactor the library to actually not need global
variables and work properly with both CommonJS/AMD as well as ES6 module
systems.

fixes #945