umd lib name remotestoragejs + xmlhttprequest external agnostic
let others solve xmlhttprequest dependency as they want when requiring
this lib. In a node environment all you'll need is to install
xmlhttprequest. in a browser environment you'll teach webpack
to use XMLHttpRequest when xmlhttprequest is required (from
webfinger.js). In case you want to include the lib with a <script>
you probably need to do something like :
`global.xmlhttprequest = XMLHttpRequest`  (to test)!