Remove UUID and MD5 utils/deps
These functions are not used in this library. Any module or package,
which depends on these hashing functions should require appropriate npm
modules itself.

In the case of the UUID dep, this actually broke node.js in production,
as the module is required in the code, but actually listed as dev
dependency (instead of normal dependency).