Link to remotestorage.io from docs
I had the experience of ending up on the doc page while I was
actually trying to go to the main website and I couldn't find any link.
Would be good to add some reference. The phrasing can be changed though.
I used the same introduction as in the readme.