Always fetch folders in the Dropbox backend
Dropbox does not offer a revision for folders,
so the revision cache used by the Dropbox backend
always returns the default value for these items.
In particular, the cached revision of a folder
never changes.

The `get` operation of the Dropbox backend
checked this revision cache to determine
whether an actual roundtrip to the Dropbox servers
is necessary.
Given what we said above, the anser to that
is always no, though.