fix default case for responseType in WireClient._fetchRequest #1151
If no responseType property is present in options parameter to WireClient._fetchRequest,
the function defaults to document and throws an error. If responseType is undefined,
the function should treat the response as text. Also, if the response is treated as text,
the syntheticXhr.responseText property should be set to processedBody in order to not
break XHR compatibility.