fix get binary files for Dropbox and GoogleDrive #1116
Dropbox and Googledrive backends first read the response as text and then tried to convert it back to raw value, however not all bytes can be encoded as text, and are lost during the conversion to text and back. Now, they request the response as ArrayBuffer and encode it to text and if they detect binary data, they return the unprocessed ArrayBuffer.

moved getTextFromArrayBuffer from wireclient to util to be used by Dropbox and GoogleDrive

fixed tests for the new change.