fix detecting changes in subfolders for Dropbox
If the root folder had subfolders, and the content of a subfolder changed remotely, the change was not picked up by the Dropbox backend.  This happened because Dropbox does not offer revision ids for folders. 
We now use RevisionCache which automatically generates revisions for folders from locally cached file revisions