fix sync related issues
revert code so when an error occurs in `fetchDelta`, emit the error as a `SyncError`. In addition, notify that the sync is done, so it can be rescheduled.

Also when hooking `sync` function, and `sync` is not yet initialized, hook `syncCycle`, which is called after sync initialization, instead of waiting for `connect` event because in doing so, a sync cycle could have already started and original `sync` is called without being hooked.