Convert Authorize to TypeScript
This fixes a bug with adding the state parameter to the redirect URI,
both in general and in nodejs. It was intended to be done only when
localStorage isn't available, but the check was incorrect, so it was
done in the opposite situation. However, it was also broken in node
(when the `btoa` function isn't available), resulting in the param not
being added there. So the test for the redirect URI didn't see the
additional param.