Port tests to Mocha and Chai
Jaribu is a bit dated now, and it's still using AMD modules.

This sets up Mocha as a test runner and Chai for assertions, as well as
ES modules/imports. A first working test file is included.

The new tests end in `.test.js` instead of `-suite.js`, and they can be
run via `npm run test:mocha` for now.