Use conda to fetch Node.js for ReadTheDocs ReadTheDocs builds are failing on the main branch due to the need for the latest Node.js. This is a temporary fix by using conda to install a more recent version of Node.js specified in the `environment.yml` file.
Even with the Node.js installed by conda, ReadTheDocs does not activate the conda environment, so it does not have immediate access to the new Node.js/npm.
Therefore, we have to manipulate the environment variable `PATH` in ``conf.py` to include the directory of the conda `bin` location.

Since `sphinx-js` does not use `npx` to run `typedoc`, we have to manually add its location to the `PATH` environment variable as well.
By default, the `typedoc` installed to the repository by `npm install` does not have executable permissions, so we have to add them in `conf.py`.

Added an auto-update for the copyright.
Added a helper script for replicating ReadTheDocs build process.
Added documentation on how to replicate ReadTheDocs build process on a local machine.